strands-sdk-tool-calling/requirements.txt
strands-agents==1.0.0
strands-agents-tools==0.2.0
ollama>=0.5.1
python-dotenv==1.0.0
